refactor: mock server classes in separate files
